+++ flutter/example/lib/main.dart
@@ -15,7 +15,7 @@ Future<void> main() async {
     (options) {
       options.dsn = _exampleDsn;
     },
-    () {
+    appRunner: () {
       // Init your App.
       runApp(MyApp());
     },